+++ src/Ext/Techno/Body.h
@@ -70,7 +70,7 @@ public:
 		void UpdateLaserTrails();
 		void InitializeLaserTrails();
 
-		virtual ~ExtData() = default;
+		virtual ~ExtData() override;
 
 		virtual void InvalidatePointer(void* ptr, bool bRemoved) override { }